sprintf() and printf() are splitted into different objects, so usage of sprintf() does not lead to linking of fonts and other heavy stuff into the rom. very useful for bgb_emu.h
